+++ tests/observer_test.php
@@ -66,7 +66,7 @@ final class observer_test extends advanced_testcase {
         $recordid = $DB->insert_record('local_deepler', $record);
 
         $event = course_updated::create(['objectid' => $course->id, 'context' => \context_course::instance($course->id)]);
-        \local_deepler\local_deepler_observer::course_updated($event);
+        observer::course_updated($event);
 
         $updatedrecord = $DB->get_record('local_deepler', ['id' => $recordid]);
         $this->assertGreaterThan($record->s_lastmodified, $updatedrecord->s_lastmodified);
@@ -96,7 +96,7 @@ final class observer_test extends advanced_testcase {
                         'other' => ['sectionnum' => $section->sectionnum ?? 1],
                 ]
         );
-        local_deepler_observer::course_section_updated($event);
+        observer::course_section_updated($event);
 
         $updatedrecord = $DB->get_record('local_deepler', ['id' => $recordid]);
         $this->assertGreaterThan($record->s_lastmodified, $updatedrecord->s_lastmodified);
@@ -129,7 +129,7 @@ final class observer_test extends advanced_testcase {
                         'instanceid' => $page->id, 'name' => 'page',
                 ],
         ]);
-        local_deepler_observer::course_module_updated($event);
+        observer::course_module_updated($event);
 
         $updatedrecord = $DB->get_record('local_deepler', ['id' => $recordid]);
         $this->assertGreaterThan($record->s_lastmodified, $updatedrecord->s_lastmodified);